+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Prefill of option elements that depend on a dynamic option element
+
 ### Added
 
-`isInfoPage` prop
+- `isInfoPage` prop
 
 ### Changed